Split tests mirroring submodule changes
